support mistle hid events sent over spi protocol by mistle fpga companion

+++ fw/ctrl_832/swap.h
@@ -1,7 +1,7 @@
 #ifndef SWAP_H
 #define SWAP_H
 
-static inline unsigned long SwapBBBB(unsigned long i) {
+static inline unsigned int SwapBBBB(unsigned int i) {
 #ifdef _M68K
     asm volatile("rol.w #8,%0\n\t"
                  "swap %0\n\t"
@@ -20,13 +20,13 @@ static inline unsigned long SwapBBBB(unsigned long i) {
 #endif
 }
 
-static inline unsigned int SwapBB(unsigned int i) {
+static inline unsigned short SwapBB(unsigned short i) {
     int result = (i >> 8) & 0xff;
     result |= (i << 8) & 0xff00;
     return (result);
 }
 
-static inline unsigned long SwapWW(unsigned long i) {
+static inline unsigned int SwapWW(unsigned int i) {
     int result = (i >> 16) & 0xffff;
     result |= (i << 16) & 0xffff0000;
     return (result);
